pyproject.toml
[tool.poetry]
name = "template-service"
version = "0.1.0"
description = ""
authors = ["NAME"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-spectacular = "^0.28.0"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
psycopg = "^3.2.3"
psycopg-binary = "^3.2.3"
jsonschema = "^4.23.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
ruff = "^0.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
